Ignore updates to non-modifiable fields instead of rejecting req
When modifying the username of a user resource or the author and
room of a message the request was rejected with an HTTP 400 Bad
Request. However when trying the API with the Web App we found out
that Ember-Data sends the whole payload when doing a PATCH
Request. This made the app unable to modify users and messages.
For this reason it was decided that any request containing un-
modifiable fields, instead of rejecting the HTTP Request we simply
ignored the changes done to the fields.