Allow tokens to be created via api
Currently tokens were not allowed to be created using the HTTP
Requests. However since it was decided that the auth should not be
a core functionality of notify, this had to change.

With this commit HTTP Consumers will be able to create tokens
without however in the future this will be restricted by GRANTS,
allowing only some users to do such requests.