Clear modifications instead of rejecting an update request
Currently when the user updated a resource having a restrict field
we were rejecting the request. However this had to be changed
since when we update a resource from EmberJS, all the fields are
sent (incl those which weren't updated). For this reason in this
commit,instead of rejecting the request, we are simply ignoring any
updates to the restricted fields.