Restrict READ CRUD operation to users with CREATE_BOT grant
Currently any user was able to request to see the access tokens
available to them, although users were only able to see tokens of
the bots they created. This commit changes this so that if a user
doesn't have the CREATE_BOT grant, when requesting to see the
access tokens a 401 Unauthorized HTTP Response is returned.